more extern cleanups
