+++ include/GX_layers.h
@@ -3,9 +3,6 @@
 
 
 #include "global.h"
-#include "GX_vramcnt.h"
-#include "error_message_reset.h"
-#include "gx.h"
 
 struct GX_LayerData
 {
@@ -18,6 +15,30 @@ typedef enum {
     GX_LAYER_TOGGLE_OFF,
     GX_LAYER_TOGGLE_ON,
 } GX_LayerToggle;
+
+
+struct GraphicsBanks
+{
+    s32 bg;
+    s32 bgextpltt;
+    s32 subbg;
+    s32 subbgextpltt;
+    s32 obj;
+    s32 objextpltt;
+    s32 subobj;
+    s32 subobjextpltt;
+    s32 tex;
+    s32 texpltt;
+};
+
+
+struct GraphicsModes {
+    u32 mode1;
+    u32 mode2;
+    u32 mode3;
+    u32 mode4;
+};
+
 
 
 THUMB_FUNC void GX_SetBanks(const struct GraphicsBanks *banks);
