+++ include/error_message_reset.h
@@ -9,30 +9,6 @@
 #include "text.h"
 
 
-
-struct GraphicsBanks
-{
-    s32 bg;
-    s32 bgextpltt;
-    s32 subbg;
-    s32 subbgextpltt;
-    s32 obj;
-    s32 objextpltt;
-    s32 subobj;
-    s32 subobjextpltt;
-    s32 tex;
-    s32 texpltt;
-};
-
-
-struct GraphicsModes {
-    u32 mode1;
-    u32 mode2;
-    u32 mode3;
-    u32 mode4;
-};
-
-
 THUMB_FUNC void VBlankHandler();
 THUMB_FUNC void PrintErrorMessageAndReset();
