+++ include/script_buffers.h
@@ -4,6 +4,7 @@
 #include "pokemon.h"
 #include "pokemon_storage_system.h"
 #include "trainer_data.h"
+#include "unk_02016B94.h"
 
 struct UnkStruct_0200AA80_sub_sub
 {
@@ -105,7 +106,7 @@ void StringExpandPlaceholders(struct ScrStrBufs * mgr, struct String * dest, str
 void ScrStrBufs_ResetBuffers(struct ScrStrBufs * mgr);
 struct UnkStruct_0200B870 * MessagePrinter_new(u32 r5, u32 r6, u32 sp4, u32 r4);
 void MessagePrinter_delete(struct UnkStruct_0200B870 * a0);
-void FUN_0200B9A8(struct UnkStruct_0200B870 * a0, int a1, int a2, int a3, int a4);
-void FUN_0200B9EC(struct UnkStruct_0200B870 * string, u32 value, u32 n, enum PrintingMode mode, int sp30, int r5, int r7);
+void FUN_0200B9A8(struct UnkStruct_0200B870 * a0, int a1, struct Window *a2, int a3, int a4);
+void FUN_0200B9EC(struct UnkStruct_0200B870 * string, u32 value, u32 n, enum PrintingMode mode, struct Window *sp30, int r5, int r7);
 
 #endif //POKEDIAMOND_SCRIPT_BUFFERS_H
